:fire: Removed use of local cache folder
Initially planned to store the JSON file of the dataset that is downloaded from online in a local folder. Dropping it for now. Will try implementing it later.

Commented visualize so it doesn't create any SVG images
