+++ benchmarking/gcn/seastar/egl_kernel.cu
@@ -21,18 +21,18 @@ extern "C" __global__ void K2
         for (; tx<feat_len; tx+=blockDim.x) {
             
             float V9_tmp = 0;
-            int offset2 = dst_id * 7 + tx;int offset3 = dst_id * 1 + tx/7;
+            int offset2 = dst_id * 1 + tx/7;int offset3 = dst_id * 7 + tx;
             
             for (int e=beg;e<end;++e) {
                 
                 int src_id = __ldg(column_indices + e);
                 int eid = __ldg(eids + e);
                 
-                int offset0 = src_id * 7 + tx;int offset1 = src_id * 1 + tx/7;
+                int offset0 = src_id * 1 + tx/7;int offset1 = src_id * 7 + tx;
                 
                 
                 
-                float V8_tmp = Vhinb[offset0]*Vnorminb[offset1];
+                float V8_tmp = Vhinb[offset1]*Vnorminb[offset0];
                 
                 
                 
@@ -48,8 +48,8 @@ extern "C" __global__ void K2
             
             
             
-            float V10_tmp = V9_tmp*Vnormcen[offset3];
-            V10[offset2] = V10_tmp;
+            float V10_tmp = V9_tmp*Vnormcen[offset2];
+            V10[offset3] = V10_tmp;
             
         }
     }
@@ -76,7 +76,7 @@ extern "C" __global__ void K2
         for (; tx<feat_len; tx+=blockDim.x) {
             
             float V14_tmp = 0;
-            int offset2 = src_id * 7 + tx;int offset3 = src_id * 1 + tx/7;
+            int offset2 = src_id * 1 + tx/7;int offset3 = src_id * 7 + tx;
             
             for (int e=beg;e<end;++e) {
                 
@@ -103,8 +103,8 @@ extern "C" __global__ void K2
             
             
             
-            float V15_tmp = V14_tmp*Vnorminb[offset3];
-            V15[offset2] = V15_tmp;
+            float V15_tmp = V14_tmp*Vnorminb[offset2];
+            V15[offset3] = V15_tmp;
             
         }
     }
