+++ benchmarking/gcn/seastar/egl_kernel.ptx
@@ -115,7 +115,7 @@
 	setp.eq.s32 	%p5, %r10, 1;
 
 $L__BB0_3:
-	.loc	1 24 55
+	.loc	1 24 25
 	mul.hi.s32 	%r31, %r65, -1840700269;
 	add.s32 	%r32, %r31, %r65;
 	shr.u32 	%r33, %r32, 31;
@@ -140,13 +140,13 @@ $L__BB0_3:
 	ld.global.nc.s32 %r36, [%rd4];
 	// end inline asm
 	.loc	1 31 29
-	mad.lo.s32 	%r37, %r35, 7, %r65;
-	.loc	1 31 59
-	add.s32 	%r38, %r35, %r15;
+	add.s32 	%r37, %r35, %r15;
+	.loc	1 31 61
+	mad.lo.s32 	%r38, %r35, 7, %r65;
 	.loc	1 35 30
-	mul.wide.s32 	%rd33, %r37, 4;
+	mul.wide.s32 	%rd33, %r38, 4;
 	add.s64 	%rd34, %rd2, %rd33;
-	mul.wide.s32 	%rd35, %r38, 4;
+	mul.wide.s32 	%rd35, %r37, 4;
 	add.s64 	%rd36, %rd1, %rd35;
 	ld.global.f32 	%f12, [%rd36];
 	ld.global.f32 	%f13, [%rd34];
@@ -167,13 +167,13 @@ $L__BB0_3:
 	ld.global.nc.s32 %r40, [%rd6];
 	// end inline asm
 	.loc	1 31 29
-	mad.lo.s32 	%r41, %r39, 7, %r65;
-	.loc	1 31 59
-	add.s32 	%r42, %r39, %r15;
+	add.s32 	%r41, %r39, %r15;
+	.loc	1 31 61
+	mad.lo.s32 	%r42, %r39, 7, %r65;
 	.loc	1 35 30
-	mul.wide.s32 	%rd39, %r41, 4;
+	mul.wide.s32 	%rd39, %r42, 4;
 	add.s64 	%rd40, %rd2, %rd39;
-	mul.wide.s32 	%rd41, %r42, 4;
+	mul.wide.s32 	%rd41, %r41, 4;
 	add.s64 	%rd42, %rd1, %rd41;
 	ld.global.f32 	%f14, [%rd42];
 	ld.global.f32 	%f15, [%rd40];
@@ -192,13 +192,13 @@ $L__BB0_3:
 	ld.global.nc.s32 %r44, [%rd8];
 	// end inline asm
 	.loc	1 31 29
-	mad.lo.s32 	%r45, %r43, 7, %r65;
-	.loc	1 31 59
-	add.s32 	%r46, %r43, %r15;
+	add.s32 	%r45, %r43, %r15;
+	.loc	1 31 61
+	mad.lo.s32 	%r46, %r43, 7, %r65;
 	.loc	1 35 30
-	mul.wide.s32 	%rd45, %r45, 4;
+	mul.wide.s32 	%rd45, %r46, 4;
 	add.s64 	%rd46, %rd2, %rd45;
-	mul.wide.s32 	%rd47, %r46, 4;
+	mul.wide.s32 	%rd47, %r45, 4;
 	add.s64 	%rd48, %rd1, %rd47;
 	ld.global.f32 	%f16, [%rd48];
 	ld.global.f32 	%f17, [%rd46];
@@ -224,13 +224,13 @@ $L__BB0_10:
 	ld.global.nc.s32 %r48, [%rd50];
 	// end inline asm
 	.loc	1 31 29
-	mad.lo.s32 	%r55, %r47, 7, %r65;
-	.loc	1 31 59
-	add.s32 	%r56, %r47, %r15;
+	add.s32 	%r55, %r47, %r15;
+	.loc	1 31 61
+	mad.lo.s32 	%r56, %r47, 7, %r65;
 	.loc	1 35 30
-	mul.wide.s32 	%rd57, %r55, 4;
+	mul.wide.s32 	%rd57, %r56, 4;
 	add.s64 	%rd58, %rd2, %rd57;
-	mul.wide.s32 	%rd59, %r56, 4;
+	mul.wide.s32 	%rd59, %r55, 4;
 	add.s64 	%rd60, %rd1, %rd59;
 	ld.global.f32 	%f18, [%rd60];
 	ld.global.f32 	%f19, [%rd58];
@@ -247,13 +247,13 @@ $L__BB0_10:
 	ld.global.nc.s32 %r50, [%rd52];
 	// end inline asm
 	.loc	1 31 29
-	mad.lo.s32 	%r57, %r49, 7, %r65;
-	.loc	1 31 59
-	add.s32 	%r58, %r49, %r15;
+	add.s32 	%r57, %r49, %r15;
+	.loc	1 31 61
+	mad.lo.s32 	%r58, %r49, 7, %r65;
 	.loc	1 35 30
-	mul.wide.s32 	%rd61, %r57, 4;
+	mul.wide.s32 	%rd61, %r58, 4;
 	add.s64 	%rd62, %rd2, %rd61;
-	mul.wide.s32 	%rd63, %r58, 4;
+	mul.wide.s32 	%rd63, %r57, 4;
 	add.s64 	%rd64, %rd1, %rd63;
 	ld.global.f32 	%f21, [%rd64];
 	ld.global.f32 	%f22, [%rd62];
@@ -270,13 +270,13 @@ $L__BB0_10:
 	ld.global.nc.s32 %r52, [%rd54];
 	// end inline asm
 	.loc	1 31 29
-	mad.lo.s32 	%r59, %r51, 7, %r65;
-	.loc	1 31 59
-	add.s32 	%r60, %r51, %r15;
+	add.s32 	%r59, %r51, %r15;
+	.loc	1 31 61
+	mad.lo.s32 	%r60, %r51, 7, %r65;
 	.loc	1 35 30
-	mul.wide.s32 	%rd65, %r59, 4;
+	mul.wide.s32 	%rd65, %r60, 4;
 	add.s64 	%rd66, %rd2, %rd65;
-	mul.wide.s32 	%rd67, %r60, 4;
+	mul.wide.s32 	%rd67, %r59, 4;
 	add.s64 	%rd68, %rd1, %rd67;
 	ld.global.f32 	%f24, [%rd68];
 	ld.global.f32 	%f25, [%rd66];
@@ -293,13 +293,13 @@ $L__BB0_10:
 	ld.global.nc.s32 %r54, [%rd56];
 	// end inline asm
 	.loc	1 31 29
-	mad.lo.s32 	%r61, %r53, 7, %r65;
-	.loc	1 31 59
-	add.s32 	%r62, %r53, %r15;
+	add.s32 	%r61, %r53, %r15;
+	.loc	1 31 61
+	mad.lo.s32 	%r62, %r53, 7, %r65;
 	.loc	1 35 30
-	mul.wide.s32 	%rd69, %r61, 4;
+	mul.wide.s32 	%rd69, %r62, 4;
 	add.s64 	%rd70, %rd2, %rd69;
-	mul.wide.s32 	%rd71, %r62, 4;
+	mul.wide.s32 	%rd71, %r61, 4;
 	add.s64 	%rd72, %rd1, %rd71;
 	ld.global.f32 	%f27, [%rd72];
 	ld.global.f32 	%f28, [%rd70];
@@ -314,14 +314,14 @@ $L__BB0_10:
 	@%p8 bra 	$L__BB0_10;
 
 $L__BB0_11:
-	.loc	1 24 55
+	.loc	1 24 25
 	add.s32 	%r63, %r15, %r2;
 	.loc	1 51 27
 	mul.wide.s32 	%rd73, %r63, 4;
 	add.s64 	%rd74, %rd16, %rd73;
 	ld.global.f32 	%f29, [%rd74];
 	mul.f32 	%f30, %f34, %f29;
-	.loc	1 24 25
+	.loc	1 24 57
 	add.s32 	%r64, %r65, %r7;
 	.loc	1 52 13
 	mul.wide.s32 	%rd75, %r64, 4;
@@ -442,7 +442,7 @@ $L__BB0_12:
 	setp.eq.s32 	%p5, %r10, 1;
 
 $L__BB1_3:
-	.loc	1 79 55
+	.loc	1 79 25
 	mul.hi.s32 	%r31, %r65, -1840700269;
 	add.s32 	%r32, %r31, %r65;
 	shr.u32 	%r33, %r32, 31;
@@ -641,14 +641,14 @@ $L__BB1_10:
 	@%p8 bra 	$L__BB1_10;
 
 $L__BB1_11:
-	.loc	1 79 55
+	.loc	1 79 25
 	add.s32 	%r63, %r15, %r2;
 	.loc	1 106 27
 	mul.wide.s32 	%rd73, %r63, 4;
 	add.s64 	%rd74, %rd16, %rd73;
 	ld.global.f32 	%f29, [%rd74];
 	mul.f32 	%f30, %f34, %f29;
-	.loc	1 79 25
+	.loc	1 79 57
 	add.s32 	%r64, %r65, %r7;
 	.loc	1 107 13
 	mul.wide.s32 	%rd75, %r64, 4;
